fix: token 유틸함수 client/server 구분해서 구현
