+++ services/web/src/features/Header/components/MyProfile/MyProfile.tsx
@@ -5,6 +5,8 @@ import { colors } from '@linker/styles';
 import clsx from 'clsx';
 import { useRouter } from 'next/navigation';
 
+import { getTokens } from '@utils/token/client';
+
 import {
   profileImage,
   profileContainer,
@@ -20,6 +22,7 @@ interface ProfileProps {
 
 function MyProfile({ isMinimize }: ProfileProps) {
   const router = useRouter();
+  const accessToken = getTokens()?.accessToken;
 
   const handleLoginClick = () => {
     if (process.env.NEXT_PUBLIC_KAKAO_LOGIN_URL == null) {
@@ -31,34 +34,41 @@ function MyProfile({ isMinimize }: ProfileProps) {
 
   return (
     <section className={clsx(profileWrapper, isMinimize && 'minimize')}>
-      <Modal>
-        <Modal.Trigger>
-          <div className={profileContainer}>
-            <Profile className={profileImage} />
+      {accessToken == null ? (
+        <Modal>
+          <Modal.Trigger>
+            <div className={profileContainer}>
+              <Profile className={profileImage} />
 
-            <div className={profileContent}>
-              <div className={profileName}>
-                <Txt typography="h5" color={colors.white}>
-                  로그인하기
+              <div className={profileContent}>
+                <div className={profileName}>
+                  <Txt typography="h5" color={colors.white}>
+                    로그인하기
+                  </Txt>
+                  <Icon name="arrow-fill-white" />
+                </div>
+
+                <Txt as="p" typography="p4" color={colors.white}>
+                  회원가입을 통해 인맥관리를 해보세요
                 </Txt>
-                <Icon name="arrow-fill-white" />
               </div>
-
-              <Txt as="p" typography="p4" color={colors.white}>
-                회원가입을 통해 인맥관리를 해보세요
-              </Txt>
             </div>
-          </div>
-        </Modal.Trigger>
+          </Modal.Trigger>
 
-        <Modal.Content className={loginModal}>
-          <div>로그인 모달</div>
+          <Modal.Content className={loginModal}>
+            <div>로그인 모달</div>
 
-          <Modal.Bottom>
-            <Button.KakaoLogin onClick={handleLoginClick} />
-          </Modal.Bottom>
-        </Modal.Content>
-      </Modal>
+            <Modal.Bottom>
+              <Button.KakaoLogin onClick={handleLoginClick} />
+            </Modal.Bottom>
+          </Modal.Content>
+        </Modal>
+      ) : (
+        <div className={profileContainer}>
+          <Profile className={profileImage} />
+          <div className={profileContent}>로그인 되어있음</div>
+        </div>
+      )}
     </section>
   );
 }
